[potter-kata] [GREEN] when_you_buy_2_different_books_you_get_5_percent_discount

+++ potter-kata/src/test/java/BookStoreTest.java
@@ -9,16 +9,16 @@ public class BookStoreTest {
     public void one_copy_of_a_book_costs_8_euros() {
         BookStore bookStore = new BookStore();
 
-        int price = bookStore.priceFor(1);
+        double price = bookStore.priceFor(1);
 
-        assertThat(price, is(8));
+        assertThat(price, is(8d));
     }
 
     @Test
     public void when_you_buy_2_different_books_you_get_5_percent_discount() {
         BookStore bookStore = new BookStore();
 
-        int price = bookStore.priceFor(1, 2);
+        double price = bookStore.priceFor(1, 2);
 
         assertThat(price, is(15.2));
     }
